p/folder moved in input fields
